Project 5: Student's Choice
A top-down game where the player moves a character on the left side of screen. While controlling the player character, there is a clone character that moves along with the player, except they move left when the player move right and vice versa. The goal of each level is to navigate the clone to the exit while avoiding enemies and solving puzzles.